extracted ImageFormatCompilerPass to abstract class for better reusability
